requirements.txt
# Web framework
streamlit>=1.29.0

# LangChain core
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13

# PDF processing
pdfplumber>=0.10.3
pypdf2>=3.0.1

# Vector database
faiss-cpu>=1.8.0

# Embedding for free mode
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
